% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyotaSiennaGasMileage.R, R/data.R
\docType{ToyotaSiennaGasMileage}
\name{ToyotaSiennaGasMileage}
\alias{ToyotaSiennaGasMileage}
\title{ToyotaSiennaGasMileage: An R package contain gas consumption data}
\format{A data frame with 8 variables:
\describe{
  \item{date}{Date, date of refueling}
  \item{gallons}{numeric, amount of fuel, in gallons}
  \item{miles}{numeric, miles driven since last refueling}
  \item{USD}{numeric, cost of fuel, in US dollars}
  \item{ethanol}{numeric, amount of ethanol, percentage}
  \item{octane}{numeric, octane rating}
  \item{notes}{character, any notes, typically brand of gas station}
  \item{inCanada}{logical, TRUE if data were collected in Canada and FALSE otherwise}
}}
\source{
<https://github.com/jarad/ToyotaSiennaGasMileage>
}
\usage{
ToyotaSiennaGasMileage
}
\description{
The ToyotaSiennaGasMileage package contains gas consumption and cost data
for a 2011 Toyota Sienna XLE. During each refuel, gas was dispensed until 
automatically stopping. At that point, the cost, amount of fuel, ethanol
content, octane, and brand were recorded. In addition, the mileage since last
refuel was recorded on a trip meter which was subsequently reset.

Each time a 2011 Toyota Sienna was refueled, the \code{date},
amount of \code{fuel}, \code{miles} driven since last refueling, \code{cost}
of the gasoline, the \code{ethanol} percentage, and the \code{octane} were
recorded.
}
\details{
The raw manually recordings are available as scans at the url below. These
data were digitized and collated to form the full data set. 

From September 2019 to July 2020, the data were recorded in Canada and
therefore the raw data are in liters (amount of fuel) and Canadian dollars
(cost). Liters were converted to gallons. Canadian dollars were converted
to US dollars using a common exchange rate (as opposed to that days exchange
rate).
}
\keyword{datasets}
