% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ToyotaSiennaGasMileage}
\alias{ToyotaSiennaGasMileage}
\title{Gas consumption data for a 2011 Toyota Sienna}
\format{A data frame with 6 variables:
\describe{
  \item{date}{date}
  \item{fuel}{amount of fuel, in gallons}
  \item{miles}{miles driven since last refueling}
  \item{cost}{cost of fuel, in US dollars}
  \item{ethanol}{amount of ethanol, percentage}
  \item{octane}{octane rating}
}}
\usage{
ToyotaSiennaGasMileage
}
\description{
Each time I refueled my 2011 Toyota Sienna, I recorded the \code{date},
amount of \code{fuel}, \code{miles} driven since last refueling, \code{cost}
of the gasoline, the \code{ethanol} percentage, and the \code{octane}.
}
\keyword{datasets}
